New updates

+++ app/FaveoStorage/Controllers/SettingsController.php
@@ -13,6 +13,9 @@ use RecursiveIteratorIterator;
 
 class SettingsController extends Controller
 {
+    public function __construct() {
+        $this->middleware(['auth','roles']);
+    }
     public function settingsIcon()
     {
         return ' <div class="col-md-2 col-sm-6">
@@ -29,31 +32,38 @@ class SettingsController extends Controller
                 </div>';
     }
 
-    public function settings()
-    {
+    public function settings() {
         try {
             $settings = new CommonSettings();
-            $directories = $this->directories();
-            $def = $settings->getOptionValue('storage', 'default');
-            $ro = $settings->getOptionValue('storage', 'root');
-            $default = 'local';
+            $default = $settings->getOptionValue('storage', 'default',true);
+            $private_folder = $settings->getOptionValue('storage', 'private-root',true);
+            $pubic_folder = $settings->getOptionValue('storage', 'public-root',true);
             $root = storage_path('app');
-            if ($def) {
-                $default = $def->option_value;
+            if (!$default) {
+                 $default = 'local';
             }
-            if ($ro) {
-                $root = $ro->option_value;
+            if (!$private_folder) {
+                $private_folder = $root . '/private';
             }
-
-            return view('storage::settings', compact('default', 'root', 'directories'));
+            if (!$pubic_folder) {
+                $pubic_folder = public_path();
+            }
+            return view('storage::settings', compact('default', 'private_folder', 'pubic_folder'));
         } catch (Exception $ex) {
             return redirect()->back()->with('fails', $ex->getMessage());
         }
     }
 
-    public function postSettings(Request $request)
-    {
+    public function postSettings(Request $request) {
         try {
+            if($request->has('private-root') && !is_dir($request->input('private-root'))){
+                $dir = $request->input('private-root');
+                throw new \Exception("'$dir'"." is not a valid directory");
+            }
+            if($request->has('private-root') && !is_writable($request->input('private-root'))){
+                $dir = $request->input('private-root');
+                throw new \Exception("'$dir'"." hasn't write permission");
+            }
             $requests = $request->except('_token');
             $this->delete();
             if (count($requests) > 0) {
@@ -63,7 +73,6 @@ class SettingsController extends Controller
                     }
                 }
             }
-
             return redirect()->back()->with('success', 'Updated');
         } catch (Exception $ex) {
             return redirect()->back()->with('fails', $ex->getMessage());
